Move ServerEndpoint & ServerEndpointRecipe to play.api.test as @ApiMayChange
Requires making play-test depend on play-netty-server.  Given play-test
already depends on play-akka-http-server, I guess this is alright.
