+++ framework/src/play-integration-test/src/test/scala/play/it/test/EndpointIntegrationSpecification.scala
@@ -7,7 +7,8 @@ package play.it.test
 import org.specs2.execute.{ AsResult, PendingUntilFixed, Result, ResultExecution }
 import org.specs2.mutable.SpecLike
 import org.specs2.specification.core.Fragment
-import play.api.test.{ ApplicationFactories, ApplicationFactory }
+
+import play.api.test.{ ApplicationFactory, ApplicationFactories, ServerEndpoint, ServerEndpointRecipe }
 
 /**
  * Mixin class for integration tests that want to run over different
